data-viz/requirements.txt
streamlit
pandas
snowflake-connector-python
cryptography
python-dotenv
plotly
